% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot.R
\name{answerPreCheckoutQuery}
\alias{answerPreCheckoutQuery}
\title{Send answers to precheckout queries}
\usage{
answerPreCheckoutQuery(pre_checkout_query_id, ok = TRUE, error_message = NULL)
}
\arguments{
\item{pre_checkout_query_id}{Unique identifier for the query to be answered.}

\item{ok}{Specify \code{TRUE} if everything is alright (goods are available, etc.) and the bot is ready to proceed with the order. Use \code{False} if there are any problems.}

\item{error_message}{(Optional). Required if \code{ok} is \code{FALSE}. Error message in human readable form that explains the reason for failure to proceed with the checkout (e.g. "Sorry, somebody just bought the last of our amazing black T-shirts while you were busy filling out your payment details. Please choose a different color or garment!"). Telegram will display this message to the user.}
}
\description{
Once the user has confirmed their payment and shipping details, the Bot API sends the final confirmation in the form of an Update with the field pre_checkout_query. Use this method to respond to such pre-checkout queries. On success, True is returned. Note: The Bot API must receive an answer within 10 seconds after the pre-checkout query was sent.
}
\details{
You can also use it's snake_case equivalent \code{answer_precheckout_query}.
}
\examples{
\dontrun{
 answerPreCheckQuery <- function(bot, update) {
  id <- update$pre_checkout_query$id
  bot$answerPreCheckoutQuery(pre_checkout_query_id = id, ok = TRUE)
}

checkout_handler <- PreCheckoutQueryHandler(answerPreCheckQuery)
}
}
