% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlinequeryhandler.R
\docType{class}
\name{InlineQueryHandler}
\alias{InlineQueryHandler}
\title{Inline queries}
\format{
An \code{\link{R6Class}} object.
}
\usage{
InlineQueryHandler(query, pattern = NULL)
}
\arguments{
\item{pattern}{(Optional). Regex pattern to test.}

\item{callback}{The callback function for this handler.
See \code{\link{Handler}} for information about this function.}
}
\description{
\code{\link{Handler}} class to handle Telegram callback queries. Optionally
based on a regex.
}
