% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot.R
\name{Invoice}
\alias{Invoice}
\title{Info about Invoice}
\usage{
Invoice(title, description, start_parameter, currency, total_amount)
}
\arguments{
\item{title}{Product name.}

\item{description}{Product description.}

\item{start_parameter}{Unique bot deep-linking parameter that can be used to generate this invoice.}

\item{currency}{Three-letter ISO 4217 currency code, see
\href{https://core.telegram.org/bots/payments#supported-currencies}.}

\item{total_amount}{Price of the product in the smallest units of the currency
(integer, not float/double). For example, for a price of US$ 1.45 pass
amount = 145. See the exp parameter.
in \href{https://core.telegram.org/bots/payments#supported-currencies}, it shows
the number of digits past the decimal point for each currency
(2 for the majority of currencies).}
}
\description{
This object contains basic information about an invoice.
}
\details{
You can also use it's snake_case equivalent \code{invoice}.
}
\examples{
\dontrun{
bot <- Bot(token = bot_token("RTelegramBot"))

bot$Invoice(
  title = 'title',
  description = 'description',
  start_parameter = 'start_parameter',
  currency = 'USD',
  total_amount = 145 # 1.45$
)
}
}
