% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheckoutqueryhandler.R
\docType{class}
\name{PreCheckoutQueryHandler}
\alias{PreCheckoutQueryHandler}
\title{Pre checkout queries}
\format{
An \code{\link{R6Class}} object.
}
\usage{
PreCheckoutQueryHandler(callback)
}
\arguments{
\item{callback}{The callback function for this handler.
See \code{\link{Handler}} for information about this function.}
}
\description{
\code{\link{Handler}} class to handle Telegram callback queries. Optionally
based on a regex.
}
