% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot.R
\name{sendInvoice}
\alias{sendInvoice}
\title{Send invoice}
\usage{
sendInvoice(
  chat_id,
  title,
  description,
  payload,
  provider_token,
  currency,
  label,
  amount,
  max_tip_amount = NULL,
  suggested_tip_amounts = NULL,
  start_parameter = NULL,
  provider_data = NULL,
  photo_url = NULL,
  photo_size = NULL,
  photo_width = NULL,
  photo_height = NULL,
  need_name = FALSE,
  need_phone_number = FALSE,
  need_email = FALSE,
  need_shipping_address = FALSE,
  send_phone_number_to_provider = FALSE,
  send_email_to_provider = FALSE,
  is_flexible = FALSE,
  disable_notification = FALSE,
  protect_content = FALSE,
  reply_to_message_id = NULL,
  allow_sending_without_reply = FALSE,
  reply_markup = NULL
)
}
\arguments{
\item{chat_id}{Unique identifier for the target chat or username of
the target channel.}

\item{title}{Product name.}

\item{description}{Product description.}

\item{payload}{Bot-defined invoice payload. This will not be displayed to
the user, use for your internal processes..}

\item{provider_token}{Payment provider token, obtained via BotFather
\href{(https://t.me/botfather)}.}

\item{currency}{Three-letter ISO 4217 currency code, see
\href{ https://core.telegram.org/bots/payments#supported-currencies}.}

\item{label}{Portion label}

\item{amount}{Price of the product in the smallest units of the currency
(integer, not float/double). For example, for a price of US$ 1.45 pass
amount = 145. See the exp parameter
in \href{https://core.telegram.org/bots/payments#supported-currencies}, it shows
the number of digits past the decimal point for each currency
(2 for the majority of currencies).}

\item{max_tip_amount}{(Optional). The maximum accepted amount for tips
in the smallest units of the currency (integer, not float/double).
For example, for a maximum tip of US$ 1.45 pass max_tip_amount = 145.
See the exp parameter
in \href{https://core.telegram.org/bots/payments#supported-currencies}, it shows
the number of digits past the decimal point for each currency
(2 for the majority of currencies). Defaults to 0.}

\item{suggested_tip_amounts}{(Optional). List of suggested amounts of tips
in the smallest units of the currency (integer, not float/double).
At most 4 suggested tip amounts can be specified. The suggested tip amounts
must be positive, passed in a strictly increased order and must not exceed
max_tip_amount.}

\item{start_parameter}{(Optional). Unique deep-linking parameter.
If left empty, forwarded copies of the sent message will have a Pay button,
allowing multiple users to pay directly from the forwarded message, using
the same invoice. If non-empty, forwarded copies of the sent message
will have a URL button with a deep link to the bot (instead of a Pay button),
 with the value used as the start parameter.}

\item{provider_data}{(Optional). Data about the invoice, which will be
shared with the payment provider. A detailed description of required fields
should be provided by the payment provider.}

\item{photo_url}{(Optional). URL of the product photo for the invoice.
Can be a photo of the goods or a marketing image for a service.
People like it better when they see what they are paying for.}

\item{photo_size}{(Optional). Photo size in bytes}

\item{photo_width}{(Optional). Photo width}

\item{photo_height}{(Optional). Photo height}

\item{need_name}{(Optional). Pass \code{TRUE}, if you require the user's full
name to complete the order}

\item{need_phone_number}{(Optional). Pass \code{TRUE}, if you require the user's
phone number to complete the order}

\item{need_email}{(Optional). Pass \code{TRUE}, if you require the user's email
address to complete the order}

\item{need_shipping_address}{(Optional). Pass \code{TRUE}, if you require
the user's shipping address to complete the order}

\item{send_phone_number_to_provider}{(Optional). Pass \code{TRUE},
if the user's phone number should be sent to provider}

\item{send_email_to_provider}{(Optional). Pass \code{TRUE},
if the user's email address should be sent to provider}

\item{is_flexible}{(Optional). Pass \code{TRUE},
if the final price depends on the shipping method}

\item{disable_notification}{(Optional). Sends the message silently.
Users will receive a notification with no sound.}

\item{protect_content}{(Optional). Protects the contents of the sent message
from forwarding and saving}

\item{reply_to_message_id}{(Optional). If the message is a reply, ID of the
original message.}

\item{allow_sending_without_reply}{(Optional). Pass \code{TRUE}, if the message
should be sent even if the specified replied-to message is not found}
}
\description{
Use this method to send invoice.
}
\details{
You can also use it's snake_case equivalent \code{send_invoice}.
}
\examples{
\dontrun{
# Just send
bot <- Bot(token = bot_token("RTelegramBot"))
chat_id <- user_id("Me")

bot$sendInvoice(
  chat_id = chat_id,
  title = 'title',
  description = 'description',
  payload = 'payload',
  provider_token = 'Your provider_token',
  currency = 'USD',
  label = 'label',
  amount = 145 # 1.45$
)

# Bot

token <- "RTelegramBot"
bot <- Bot(token = token)
updater <- Updater(token)

Pay <- function(bot, update){
  bot$sendInvoice(
    chat_id = update$message$chat_id,
    title = 'title',
    description = 'description',
    payload = 'payload',
    provider_token = 'Your provider_token',
    currency = 'USD',
    label = 'label',
    amount = 145 # 1.45$
  )
}

pay_cmd     <- CommandHandler('pay', Pay)


answerQuery <- function(bot, update) {
  chat_id <- update$pre_checkout_query$id
  bot$answerPreCheckoutQuery(chat_id)
}

pay_cmd_handler     <- CommandHandler('pay', Pay)
checkout_handler <- PreCheckoutQueryHandler(answerQuery)

updater <- updater +
  pay_cmd_handler + checkout_handler

updater$start_polling()

}
}
